+++ crates/swc_css_minifier/src/compressor/media.rs
@@ -1,9 +1,10 @@
 use std::mem::take;
 
-use swc_common::{EqIgnoreSpan, DUMMY_SP};
+use swc_common::DUMMY_SP;
 use swc_css_ast::*;
 
 use super::Compressor;
+use crate::util::dedup;
 
 impl Compressor {
     fn is_first_media_in_parens(&self, media_condition: &MediaCondition) -> bool {
@@ -29,19 +30,7 @@ impl Compressor {
     }
 
     pub(super) fn compress_media_query_list(&mut self, media_query_list: &mut MediaQueryList) {
-        let mut already_seen: Vec<MediaQuery> = Vec::with_capacity(media_query_list.queries.len());
-
-        media_query_list.queries.retain(|children| {
-            for already_seen_complex_selector in &already_seen {
-                if already_seen_complex_selector.eq_ignore_span(children) {
-                    return false;
-                }
-            }
-
-            already_seen.push(children.clone());
-
-            true
-        });
+        dedup(&mut media_query_list.queries);
     }
 
     pub(super) fn compress_media_condition(&mut self, n: &mut MediaCondition) {
