+++ crates/swc_css_minifier/src/compressor/selector.rs
@@ -1,10 +1,9 @@
-use std::mem::take;
-
 use swc_atoms::js_word;
-use swc_common::{EqIgnoreSpan, DUMMY_SP};
+use swc_common::DUMMY_SP;
 use swc_css_ast::*;
 
 use super::Compressor;
+use crate::util::dedup;
 
 impl Compressor {
     pub(super) fn comrpess_selector_list(&mut self, selector_list: &mut SelectorList) {
@@ -306,37 +305,4 @@ impl Compressor {
             }));
         }
     }
-}
-
-fn dedup<T>(v: &mut Vec<T>)
-where
-    T: EqIgnoreSpan,
-{
-    let mut remove_list = vec![];
-
-    for (i, i1) in v.iter().enumerate() {
-        for (j, j1) in v.iter().enumerate() {
-            if i < j && i1.eq_ignore_span(j1) {
-                remove_list.push(j);
-            }
-        }
-    }
-    // Fast path. We don't face real duplciate in most cases.
-    if remove_list.is_empty() {
-        return;
-    }
-
-    let new = take(v)
-        .into_iter()
-        .enumerate()
-        .filter_map(|(idx, value)| {
-            if remove_list.contains(&idx) {
-                None
-            } else {
-                Some(value)
-            }
-        })
-        .collect::<Vec<_>>();
-
-    *v = new;
 }
